[split globel]
