+++ src/App.tsx
@@ -1,5 +1,6 @@
 // import React from 'react'
-import GlobeComponent from './GlobeComponent'
+// import GlobeComponent from './GlobeComponent'
+import GlobeComponent from "./globe/GlobeComponent"
 
 function App() {
   return <GlobeComponent />
