+++ src/types.ts
@@ -1,0 +1,8 @@
+export type News = {
+  id: string;
+  title: string;
+  content: string;
+  time: string;
+  lat: number;
+  lon: number;
+};
